# Feature/aruco speedup