--- a/modules/aruco/samples/detect_markers.cpp
+++ b/modules/aruco/samples/detect_markers.cpp
@@ -59,7 +59,8 @@ const char* keys  =
         "{dp       |       | File of marker detector parameters }"
         "{r        |       | show rejected candidates too }"
         "{refine   |       | Corner refinement: CORNER_REFINE_NONE=0, CORNER_REFINE_SUBPIX=1,"
-        "CORNER_REFINE_CONTOUR=2, CORNER_REFINE_APRILTAG=3}";
+        "CORNER_REFINE_CONTOUR=2, CORNER_REFINE_APRILTAG=3}"
+        "{ar3vid   |       | Adapt the paramater tau_i if aruco3 functionality is used. }";
 }
 
 /**
@@ -101,6 +102,12 @@ static bool readDetectorParameters(string filename, Ptr<aruco::DetectorParameter
     fs["maxErroneousBitsInBorderRate"] >> params->maxErroneousBitsInBorderRate;
     fs["minOtsuStdDev"] >> params->minOtsuStdDev;
     fs["errorCorrectionRate"] >> params->errorCorrectionRate;
+    // new aruco functionality
+    fs["useAruco3Detection"] >> params->useAruco3Detection;
+    fs["minSideLengthCanonicalImg"] >> params->minSideLengthCanonicalImg;
+    fs["minMarkerLengthRatioOriginalImg"] >> params->minMarkerLengthRatioOriginalImg;
+    fs["cameraMotionSpeed"] >> params->cameraMotionSpeed;
+    fs["useGlobalThreshold"] >> params->useGlobalThreshold;
     return true;
 }
 
@@ -121,6 +128,7 @@ int main(int argc, char *argv[]) {
     bool showRejected = parser.has("r");
     bool estimatePose = parser.has("c");
     float markerLength = parser.get<float>("l");
+    bool useAruco3DynamicUpdates = parser.has("ar3vid");
 
     Ptr<aruco::DetectorParameters> detectorParams = aruco::DetectorParameters::create();
     if(parser.has("dp")) {
@@ -165,7 +173,7 @@ int main(int argc, char *argv[]) {
     int waitTime;
     if(!video.empty()) {
         inputVideo.open(video);
-        waitTime = 0;
+        waitTime = 1;
     } else {
         inputVideo.open(camId);
         waitTime = 10;
@@ -173,7 +181,8 @@ int main(int argc, char *argv[]) {
 
     double totalTime = 0;
     int totalIterations = 0;
-
+    float new_marker_length_ratio = 0.0;
+    size_t total_nr_detected_corners = 0;
     while(inputVideo.grab()) {
         Mat image, imageCopy;
         inputVideo.retrieve(image);
@@ -185,7 +194,13 @@ int main(int argc, char *argv[]) {
         vector< Vec3d > rvecs, tvecs;
 
         // detect markers and estimate pose
-        aruco::detectMarkers(image, dictionary, corners, ids, detectorParams, rejected);
+        if (useAruco3DynamicUpdates) {
+            // if new aruco3 features are used, we can also set the new min
+            // marker length ratio dymamically from the last frame
+            detectorParams->minMarkerLengthRatioOriginalImg = new_marker_length_ratio;
+        }
+        new_marker_length_ratio = aruco::detectMarkers(image, dictionary, corners, ids, detectorParams, rejected);
+        total_nr_detected_corners += ids.size();
         if(estimatePose && ids.size() > 0)
             aruco::estimatePoseSingleMarkers(corners, markerLength, camMatrix, distCoeffs, rvecs,
                                              tvecs);
@@ -217,6 +232,8 @@ int main(int argc, char *argv[]) {
         char key = (char)waitKey(waitTime);
         if(key == 27) break;
     }
+
+    cout<<"Total number detected corners: "<<total_nr_detected_corners<<"\n";
 
     return 0;
 }