--- a/modules/aruco/src/aruco.cpp
+++ b/modules/aruco/src/aruco.cpp
@@ -1107,7 +1107,7 @@ float detectMarkers(InputArray _image, const Ptr<Dictionary> &_dictionary, Outpu
 
     /// Step 0: equation (2) from paper [1]
     const int tau_i_dot = _params->minSideLengthCanonicalImg +
-            std::max(grey.cols, grey.rows) * _params->minMarkerLengthRatioOriginalImg;
+            (int)((float)std::max(grey.cols, grey.rows) * _params->minMarkerLengthRatioOriginalImg);
 
     //// Step 0.1: resize image with equation (1) from paper [1]
     float fxfy = (float)_params->minSideLengthCanonicalImg / tau_i_dot;