… modules/aruco/samples/detector_params.yml
… modules/aruco/test/test_boarddetection.cpp