### Pull Request Readiness Checklist

See details at https://github.com/opencv/opencv/wiki/How_to_contribute#making-a-good-pull-request

- [x] I agree to contribute to the project under Apache 2 License.
- [x] To the best of my knowledge, the proposed patch is not based on a code under GPL or other license that is incompatible with OpenCV
- [x] The PR is proposed to proper branch
- [x] There is reference to original bug report and related work
- [x] There is accuracy test, performance test and test data in opencv_extra repository, if applicable
      Patch to opencv_extra has the same branch name.
- [x] The feature is well documented and sample code can be built with the project CMake

Hi,

I improved the Aruco module (I will call it Aruco3 as the authors did). It implements the improvements that were proposed in the paper: "Speeded up detection of squared fiducial markers", 2018, FJ Romera-Ramirez et al.

The implementation is not entirely finished, meaning that I would like to know from you guys where I should add tests and so on. 
I uploaded simple test videos: https://drive.google.com/drive/folders/153UTdWMOwxIsn2nxdWW0itV8na4AANZH?usp=sharing
The folder contains 3 videos with 4k, 1k and VGA resolution to test the speed improvements and detection rates. However are there any particular tests for the Aruco module to compare against the original implementation?

I added the functionality to the example_aruco_detect_markers.cpp file. In addition, I added new flags to the detector_params.yml. If Aruco3 is turned off, the code should behave as the original Aruco implementation.
```yaml
# new aruco functionality
useAruco3Detection: 0
minSideLengthCanonicalImg: 32 # 16, 32, 64 --> tau_c from the paper
minMarkerLengthRatioOriginalImg: 0.08 # range [0,0.2] --> tau_i from the paper
cameraMotionSpeed: 0.5 # range [0,1) --> tau_s from the paper
useGlobalThreshold: 0
```
The parameters are explained in the documentation and reference to equations and sections from the paper.

Some preliminary results for #detected and mean computation per image in [ms]:
Call: 
``` bash
./example_aruco_detect_markers -v=4k_test.mp4 -d=16 -dp=detector_params.yml -refine=1
```
For all datasets: tau_c = 32pixel
| Dataset | Aruco | Aruco3 (tau_i = 0.1)  | Aruco3 (tau_i = 0.05)  | Aruco3 (tau_i = 0.02)  |
|--|--|--|--|--|
| VGA | 921 / 2.91ms |  704 / 0.95ms |  761 / 1.1ms  | 909 / 2.1ms |
| 1080p |  1186 / 14ms |  995 / 2.6ms| 1100 / 3.1ms | 1180 / 5.3ms |
| 4K | 1013 / 63ms | 931 / 7.6ms |  1011 / 8.1ms  |  1071 / 13.3ms | 

For all datasets: tau_c = 16pixel
| Dataset | Aruco | Aruco3 (tau_i = 0.1)  | Aruco3 (tau_i = 0.05)  | Aruco3 (tau_i = 0.02)  |
|--|--|--|--|--|
| VGA |  921 / 2.91ms  | 474 / 0.8ms | 796 / 0.9ms |  919 / 1.7ms |
| 1080p | 1186 / 14ms  | 132 / 2.9ms  | 1086 / 2.9 ms  | 1173 / 4.0 ms | 
| 4K | 1013 / 63ms | 118 / 7.0ms |  1013 /7.4ms  |  1052 / 8.3ms | 

Now instead of fixing tau_i we can also let the algorithm decide it by using the marker with the smallest perimeter from the last frame. This only works for video streams of course. In addition, we can use global thresholding instead of adaptive thresholding, assuming that the illuminations does not change much. 
Call: 
``` bash
./example_aruco_detect_markers -v=4k_test.mp4 -d=16 -dp=detector_params.yml -refine=1 -ar3vid
```

For all datasets: tau_c = 32pixel, tau_s = 0.1, useGlobalThreshold=1
| Dataset | Aruco | Aruco3 |
|--|--|--|
| VGA |  921 / 2.91ms  | 758 / 1.0ms  |
| 1080p | 1186 / 14ms  | 1174 / 2.9ms  | 
| 4K | 1013 / 63ms | 1036 / 8.8ms | 

For all datasets: tau_c = 16pixel, tau_s = 0.1, useGlobalThreshold=1
| Dataset | Aruco | Aruco3 |
|--|--|--|
| VGA |  921 / 2.91ms  | 905 / 0.7ms  |
| 1080p | 1186 / 14ms  | 1156 / 2.8ms  | 
| 4K | 1013 / 63ms | 1022 / 7.7ms | 

Let me know what you think.
Cheers
Steffen