@urbste 2020-12-17
Ok I can see that already some Aruco tests are failing. I will have a look into it.