@urbste 2021-02-10
I added tests. I also fixed some small bugs. Can someone check why the build checks are failing? When I click on details it looks like all tests run successfully.