@alalek 2021-02-10
You can see compilation warnings [here](http://pullrequest.opencv.org/#/summary/contrib) (should be resolved before merge).