@urbste 2021-08-03
I am still having it in the back of my head. However, the requested changes seem to be quite significant and I did not find the time to look at it. We were using this branch in prototype, however, we moved away from it. So if someone likes to take over please go ahead!