@urbste 2021-08-04
@joellyons no we did not have problems with it. We wanted to use it in WebAssembly running on a smartphone to track the pose in realtime. However our board had a lot of markers and we never reached >30fps on Android devices. So we kind of changed our algorithm and now only occasionally update the pose using the ARuco board. The Aruco3 algorithm however really shines if you want to track the board over time, so you can update the tau_i parameter and use the global otsu thesholding. So we decided to stick with the classic Aruco implementation.

Still with my PR it is possible to use the classic impementation. I will see if I find the time to finish this PR :/