@joellyons 2021-08-04
@urbste thank you so much for taking the time to explain. We are wanting to track a small board over time, so this sounds perfect. But, I agree it's hard to find extra time to work on open source projects like this.